chore : move BENCHTTP to shared
